Alpha 1.0.1
Added an interface for the Character, Charcter class inherits from the interface.

Also added a list for player characters, the player can add new items to this list.

Also added a Character creator state, which inherits from the base GameStates class. Additional methods to be implemented at a later given time.

Added a Character class that inherits from the interface, this class makes the other properties accessible for future reference.